Ajuste do projeto final

+++ Prof André/Projeto_Final_1/crud/src/main/java/crud/projeto1/crud/Controllers/HospedeREST.java
@@ -1,7 +1,10 @@
 package crud.projeto1.crud.Controllers;
 
+import crud.projeto1.crud.Entities.Veiculo;
 import crud.projeto1.crud.Repositories.RepositorioHospede;
 import crud.projeto1.crud.Entities.Hospede;
+import crud.projeto1.crud.Repositories.RepositorioVeiculo;
+import crud.projeto1.crud.ViewModels.HospedeFromRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,6 +21,9 @@ public class HospedeREST {
     @Autowired
     private RepositorioHospede repositorio;
 
+    @Autowired
+    private RepositorioVeiculo repositorio_veiculo;
+
     @GetMapping
     public List<Hospede> listar(){
 
@@ -26,10 +32,14 @@ public class HospedeREST {
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void salvar(@RequestBody Hospede hospedeFromRequest){
+    public void salvar(@RequestBody HospedeFromRequest hospedeFromRequest){
 
         Hospede hospede = new Hospede(hospedeFromRequest.getNome_Completo());
         repositorio.save(hospede);
+
+        Veiculo veiculo = new Veiculo(hospedeFromRequest.getPlaca(), hospede);
+        repositorio_veiculo.save(veiculo);
+
         log.info("Hospede cadastrado com sucesso");
     }
 
@@ -42,12 +52,23 @@ public class HospedeREST {
         if(result.isPresent()){
             repositorio.save(hospede);
         }
+
         log.info("Hospede alterado com sucesso!!");
     }
 
     @DeleteMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void excluir(@PathVariable int id){
+
+        Optional<Hospede> hospede_optional = repositorio.findById(id);
+        if(hospede_optional.isEmpty()) return;
+
+        Hospede hospede = hospede_optional.get();
+
+        if(hospede.getVeiculo() != null){
+
+            repositorio_veiculo.deleteById(hospede.getVeiculo().getId());
+        }
 
         repositorio.deleteById(id);
         log.info("Hospede deletado com sucesso!!");
